refactor: move TasksEmptyScreen to composable file
