+++ app/src/main/java/com/baghdad/tudee/ui/composable/TasksEmptyScreen.kt
@@ -1,4 +1,4 @@
-package com.baghdad.tudee.ui.screens.tasks.components
+package com.baghdad.tudee.ui.composable
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
